==> R/02_illustrative_data.R <==
ill_nwp <- list(
    ay = c("1 2 3 4 5 6 7 8 9 10"),
    prem = c("5812 4908 5454 5165 5214 5230 4992 5466 5226 4962")
) %>% 
    purrr::map(~ stringr::str_split(., pattern = " ")) %>%
    tibble::as_tibble() %>%
    tidyr::unnest()

raw_netincloss <- c(
    "1988 1722 3830 3603 3835 3873 3895 3918 3918 3917 3917 1997",
    "1989 1581 2192 2528 2533 2528 2530 2534 2541 2538 2532 1998",
    "1990 1834 3009 3488 4000 4105 4087 4112 4170 4271 4279 1999",
    "1991 2305 3473 3713 4018 4295 4334 4343 4340 4342 4341 2000",
    "1992 1832 2625 3086 3493 3521 3563 3542 3541 3541 3587 2001",
    "1993 2289 3160 3154 3204 3190 3206 3351 3289 3267 3268 2002",
    "1994 2881 4254 4841 5176 5551 5689 5683 5688 5684 5684 2003",
    "1995 2489 2956 3382 3755 4148 4123 4126 4127 4128 4128 2004",
    "1996 2541 3307 3789 3973 4031 4157 4143 4142 4144 4144 2005",
    "1997 2203 2934 3608 3977 4040 4121 4147 4155 4183 4181 2006")

parse_row <- purrr::as_mapper(~ stringr::str_split(., pattern = " ") %>% 
                                  unlist() %>%
                                  { 
                                      tibble::tibble(
                                          ay = .[1], 
                                          source = .[length(.)],
                                          lag = seq_len(length(.) - 2),
                                          vals = as.numeric(.[-c(1, length(.))])
                                          )
                                  }
                              )

ill_netincloss <- raw_netincloss %>%
    purrr::map_df(parse_row)


raw_netpdloss <- c(
    "1988 1722 3830 3603 3835 3873 3895 3918 3918 3917 3917 1997",
    "1989 1581 2192 2528 2533 2528 2530 2534 2541 2538 2532 1998",
    "1990 1834 3009 3488 4000 4105 4087 4112 4170 4271 4279 1999",
    "1991 2305 3473 3713 4018 4295 4334 4343 4340 4342 4341 2000",
    "1992 1832 2625 3086 3493 3521 3563 3542 3541 3541 3587 2001",
    "1993 2289 3160 3154 3204 3190 3206 3351 3289 3267 3268 2002",
    "1994 2881 4254 4841 5176 5551 5689 5683 5688 5684 5684 2003",
    "1995 2489 2956 3382 3755 4148 4123 4126 4127 4128 4128 2004",
    "1996 2541 3307 3789 3973 4031 4157 4143 4142 4144 4144 2005",
    "1997 2203 2934 3608 3977 4040 4121 4147 4155 4183 4181 2006")

ill_netpdloss <- raw_netpdloss %>% 
    purrr::map_df(parse_row)

saveRDS(ill_nwp, "cache/ill_nwp.rds")
saveRDS(ill_netincloss, "cache/ill_incloss.rds")
saveRDS(ill_netpdloss, "cache/ill_pdloss.rds")
