==> R/02_selected_groups.R <==
library(tidyverse)

select_grcode <- c("353 13420 353 13641 86 13528 620 14370",
  "388 13439 388 13889 337 14176 669 14915",
  "620 13641 620 14044 353 14320 671 15113",
  "833 13889 692 14176 388 14508 683 15148",
  "1066 14044 715 14257 671 14974 715 15210",
  "1090 14176 1066 14311 715 15148 833 15571",
  "1538 14257 1090 14443 1252 15199 1538 17043",
  "1767 14320 1538 15199 1538 15334 1767 17450",
  "2135 14974 1767 15393 1767 18309 2003 17493",
  "2208 15199 2003 15660 2135 18538 2135 18163",
  "2623 18163 2143 16373 2712 18767 2143 18686",
  "2712 18767 3240 16799 3034 18791 2208 24830",
  "3240 19020 4839 18163 3240 21172 3240 26797",
  "3492 21270 5185 18791 5185 23108 5185 27065",
  "4839 25275 6947 23574 6408 26433 5320 28550",
  "5185 27022 7080 25275 6807 27529 6459 30449",
  "6408 27065 8427 27022 7080 30589 6947 30651",
  "6459 29440 8559 27065 8559 32875 7625 32301",
  "6947 31550 10022 27499 8672 33499 10657 33049",
  "7080 32301 13420 27766 9466 34576 13501 36315",
  "8427 34606 13439 29440 10385 35408 13919 38733",
  "10022 35483 13501 31550 10699 37370 13994 41068",
  "10308 37036 13528 34509 11347 38687 14044 41580",
  "11037 38733 13587 34592 11703 38733 14176 42439",
  "11118 44598 13595 34606 13439 41300 14257 43354")

selected_groups <- select_grcode %>% 
    str_split(pattern = " ") %>%
    map_dfr( ~tibble(comauto = .x[1:2], 
                     ppauto = .x[3:4], 
                     wkcomp = .x[5:6], 
                     othliab = .x[7:8])
    ) %>%
    gather(key = lob, value = grcode) %>%
    mutate(grcode = as.numeric(grcode))

cache_dir <- glue::glue("{here::here()}/cache")
if (!dir.exists(cache_dir)) dir.create(cache_dir)

saveRDS(selected_groups, file = fs::path(cache_dir, "selected_groups.rds"))
