==> R/04_validation_mack_model.Rmd <==
---
title: "Validating Mack Method"
output: html_notebook
---

Reproducing the validation of the Mack method from Meyees page 8 and 9.

```{r}
suppressPackageStartupMessages({
    library(tidyverse)
    library(fs)
})

cache_path <- fs::path(here::here(), "cache")
selected_groups <- readRDS(file =  fs::path(cache_path, "selected_groups.rds"))
comauto <- readr::read_csv(fs::path(cache_path, "comauto_pos.csv"))
```


```{r}
path_to_utils <- glue::glue("{here::here()}/R/00_utils.R")
source(path_to_utils)
comm_auto_groups <- selected_groups %>% 
    filter(lob == "commauto")

suppressWarnings({
    comm_auto_mack <- comauto %>%
        filter(GRCODE %in% comm_auto_groups$grcode) %>%
        split(.$GRCODE) %>%
        purrr::map(purrr::safely(calc_mack_estimates))
})

check_work <- comm_auto_mack %>% 
    map_dbl( ~ is.null(.x$error)) %>% 
    sum()

results_to_test <- map_dfr(comm_auto_mack, ~ .x$result)
```

I found `r check_work` out of 50 to be correct.

```{r}

pp_plot_data <- select(results_to_test, percentile) %>% 
    arrange(percentile) %>%
    mutate(exp_percentile = row_number() / (n() + 1), unifdist = row_number() / n ())
           
ggplot(pp_plot_data, aes(x = exp_percentile, y = percentile)) +
    geom_point() + 
    geom_abline(slope = 1, intercept = 0, color = "blue") +
    geom_abline(slope = 1, intercept = 136 / sqrt(50) * 0.01, color = "blue") +
    geom_abline(slope = 1, intercept = -136 / sqrt(50) * 0.01, color = "blue")
```

Kolmogorov–Smirnov (K–S) Test

```{r}
ks.test(pp_plot_data$percentile, "punif")
```
